requirements.txt
python-docx
scikit-learn
pandas
spacy
joblib
